[FIX] Pages couldn't leave fora [FIX] Join button redirects correctly

+++ app/helpers/forums_helper.rb
@@ -12,16 +12,19 @@ module ForumsHelper
     end
 
     divided = false
-    if active_for_user?(:notifications, current_user)
-      divided = true
-      if current_profile.following?(@forum)
-        link_items << link_item(t('forums.unfollow'), follows_path(forum_id: @forum.web_url), fa: 'times', divider: 'top', data: {method: 'delete', 'skip-pjax' => 'true'})
-      else
-        link_items << link_item(t('forums.follow'), follows_path(forum_id: @forum.web_url), fa: 'check', divider: 'top', data: {method: 'create', 'skip-pjax' => 'true'})
+    if policy(@forum).is_member?
+      if active_for_user?(:notifications, current_user)
+        divided = true
+        if current_profile.following?(@forum)
+          link_items << link_item(t('forums.unfollow'), follows_path(forum_id: @forum.web_url), fa: 'times', divider: 'top', data: {method: 'delete', 'skip-pjax' => 'true'})
+        else
+          link_items << link_item(t('forums.follow'), follows_path(forum_id: @forum.web_url), fa: 'check', divider: 'top', data: {method: 'create', 'skip-pjax' => 'true'})
+        end
       end
+      link_items << link_item(t('forums.leave'), forum_membership_path(@forum.web_url, current_profile), fa: 'sign-out', divider: (divided ? 'top' : nil),
+                              data: {method: :delete, 'skip-pjax' => 'true', confirm: t('forums.leave_confirmation')})
     end
-    link_items << link_item(t('forums.leave'), forum_membership_path(@forum.web_url, current_profile), fa: 'sign-out', divider: (divided ? 'top' : nil), data: {method: :delete, 'skip-pjax' => 'true', confirm: t('forums.leave_confirmation')}) if policy(@forum).is_member?
-    link_items << link_item(t('forums.discover'), forums_url, fa: 'compass') if policy(@forum).is_member?
+    link_items << link_item(t('forums.discover'), forums_url, fa: 'compass')
     dropdown_options(resource.name, [{items: link_items}], triggerTag: 'h1', fa_after: 'fa-angle-down')
   end
